refactor(api): move certification center accept invitation route
to src/team
