+++ api/lib/application/certification-center-invitations/index.js
@@ -7,30 +7,6 @@ import { certificationCenterInvitationController } from './certification-center-
 const register = async function (server) {
   server.route([
     {
-      method: 'POST',
-      path: '/api/certification-center-invitations/{id}/accept',
-      config: {
-        auth: false,
-        handler: certificationCenterInvitationController.acceptCertificationCenterInvitation,
-        validate: {
-          params: Joi.object({
-            id: identifiersType.certificationCenterInvitationId,
-          }),
-          payload: Joi.object({
-            data: {
-              id: Joi.string().required(),
-              type: Joi.string().required(),
-              attributes: {
-                code: Joi.string().required(),
-                email: Joi.string().email().required(),
-              },
-            },
-          }),
-        },
-        tags: ['api'],
-      },
-    },
-    {
       method: 'GET',
       path: '/api/certification-center-invitations/{id}',
       config: {
